moved `graph/_input_check.py` to `graph/utils`
