+++ pyrigi/framework/base.py
@@ -11,7 +11,7 @@ import numpy as np
 import sympy as sp
 from sympy import Matrix
 
-import pyrigi.graph._input_check as _graph_input_check
+import pyrigi.graph.utils._input_check as _graph_input_check
 import pyrigi.misc._input_check as _input_check
 from pyrigi.data_type import (
     Edge,
