+++ pyrigi/graph/constructions/extensions.py
@@ -9,7 +9,7 @@ from typing import Iterable
 
 import networkx as nx
 
-import pyrigi.graph._input_check as _graph_input_check
+import pyrigi.graph.utils._input_check as _graph_input_check
 import pyrigi.misc._input_check as _input_check
 from pyrigi.data_type import Edge, Sequence, Vertex
 from pyrigi.exception import NotSupportedValueError
