+++ pyrigi/graph/graph.py
@@ -20,7 +20,7 @@ from pyrigi.misc.misc import _doc_category as doc_category
 from pyrigi.misc.misc import _generate_category_tables
 from pyrigi.plot_style import PlotStyle
 
-from . import _input_check as _graph_input_check
+from .utils import _input_check as _graph_input_check
 from . import general
 from .assorted import apex, separating_set
 from .constructions import extensions
