+++ pyrigi/graph/rigidity/generic.py
@@ -9,7 +9,7 @@ from typing import TypeVar
 import networkx as nx
 from sympy import oo
 
-import pyrigi.graph._input_check as _graph_input_check
+import pyrigi.graph.utils._input_check as _graph_input_check
 import pyrigi.graph.constructions.extensions as graph_extension
 import pyrigi.graph.sparsity.sparsity as sparsity
 import pyrigi.misc._input_check as _input_check
