+++ pyrigi/graph/utils/_input_check.py
