+++ test/graph/utils/test__input_check.py
@@ -1,6 +1,6 @@
 import pytest
 
-import pyrigi.graph._input_check as _graph_input_check
+import pyrigi.graph.utils._input_check as _graph_input_check
 import pyrigi.graphDB as graphs
 from pyrigi.exception import LoopError
 from pyrigi.graph import Graph
